feat: 优化配置Dialog
